added jpa adapter with address embedable and repository + docker file and scmea file for database tables, still wondering how to do the relationship between owner and restaurant in the domain and database

+++ src/main/java/be/kdg/ivanov_kaloyan_prog6_backend/restaurant/domain/Dish.java
@@ -1,5 +1,6 @@
 package be.kdg.ivanov_kaloyan_prog6_backend.restaurant.domain;
 
+import java.math.BigDecimal;
 import java.util.*;
 
 public class Dish {
@@ -15,7 +16,7 @@ public class Dish {
 
    private String description;
 
-   private Double price;
+   private BigDecimal price;
 
    private String pictureURL;
 
@@ -24,7 +25,7 @@ public class Dish {
    }
 
     public Dish(String name, DishType type, String description,
-                Double price, String pictureURL) {
+                BigDecimal price, String pictureURL) {
         this.dishId = DishId.create();
         this.name = name;
         this.type = type;
